refactor emulators into package
blacking
